added test

+++ ecommerce/src/main/java/com/lala/ecommerce/service/impl/CategoryServiceImpl.java
@@ -1,9 +1,11 @@
 package com.lala.ecommerce.service.impl;
 
+import com.lala.ecommerce.exception.CategoryDeleteException;
 import com.lala.ecommerce.exception.CategoryDuplicateException;
 import com.lala.ecommerce.exception.CategoryNotFoundException;
 import com.lala.ecommerce.model.Category;
 import com.lala.ecommerce.repository.CategoryRepository;
+import com.lala.ecommerce.repository.ProductRepository;
 import com.lala.ecommerce.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,7 @@ import java.util.Optional;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final ProductRepository productRepository;
 
     @Override
     public Category createCategory(Category category) {
@@ -29,6 +32,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Long id) {
+        Long productCountOfCategory = productRepository.getProductCountByCategoryId(id);
+        if(productCountOfCategory > 0) {
+            throw new CategoryDeleteException("you can not delete this category because category has " + productCountOfCategory + " product");
+        }
         categoryRepository.deleteById(id);
     }
